iniciado desenvolvimento do PJBL2

+++ codes/src/PJBL1/Chave.java
@@ -1,11 +1,11 @@
 package PJBL1;
 
 public class Chave {
-    private byte[] meuArray;
-    private byte index;
+    private int[] meuArray;
+    private int index;
 
     public Chave(){
-        this.meuArray = new byte[100];
+        this.meuArray = new int[100];
         this.index = 0;
         // cria o array e preenche todos os espacos com 0
         for(byte i=0;i<this.meuArray.length;i++){
@@ -17,7 +17,7 @@ public class Chave {
         this.meuArray[index] = e;
     }
 
-    public byte[] getArray(){
+    public int[] getArray(){
         return this.meuArray;
     }
